requirements.txt

fastapi==0.111.0
uvicorn==0.30.1
sqlalchemy==2.0.32
pydantic==2.8.2
pydantic-settings==2.4.0
python-jose==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
requests==2.32.3
cryptography==43.0.1
Pillow==10.4.0
